requirements.txt
requests==2.31.0
pandas==1.5.3
openai==0.27.8
langchain==0.1.12
python-dotenv==1.0.0
chromadb==0.4.3
tiktoken==0.4.0